semantic markup for main events listing
Notably made each month’s list of events an unordered list. Each month
is its own <section> with @aria-labelledby referencing the @id of the
child <h2>. That @id is the name of the month, which should be fine as
long as we don’t list events more than 12 months in advance.
Also cleaned up extraneous attributes on <main>.